# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoBomb API
  version: 1.0.0
  description: |
    Production-ready photo service API supporting uploads, metadata management, 
    search, albums, and sharing with presigned URLs and async processing.
  contact:
    name: PhotoBomb Team
    email: api@photobomb.example.com
servers:
  - url: https://api.photobomb.example.com/api/v1
    description: Production
  - url: https://api-staging.photobomb.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /upload/presign:
    post:
      tags: [upload]
      summary: Get presigned upload URL
      description: |
        Request a presigned URL for direct browser-to-B2 upload.
        Includes duplicate detection based on SHA256.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [filename, size_bytes, mime_type, sha256]
              properties:
                filename:
                  type: string
                  example: "IMG_1234.jpg"
                size_bytes:
                  type: integer
                  example: 5242880
                  description: File size in bytes
                mime_type:
                  type: string
                  example: "image/jpeg"
                  enum: [image/jpeg, image/png, image/heic, image/webp, image/avif]
                sha256:
                  type: string
                  example: "a3c8f9e7b2d4..."
                  description: Hex-encoded SHA256 hash (computed client-side)
      responses:
        '200':
          description: Presigned URL generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_id:
                    type: string
                    format: uuid
                  presigned_url:
                    type: string
                    format: uri
                    description: B2 presigned upload URL (15min expiry)
                  multipart_chunk_size:
                    type: integer
                    example: 5242880
                    description: Recommended chunk size for multipart upload
                  expires_at:
                    type: string
                    format: date-time
        '409':
          description: Duplicate photo detected
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "duplicate"
                  photo_id:
                    type: string
                    format: uuid
                  message:
                    type: string
                    example: "This photo was already uploaded on 2024-01-15"
  /upload/confirm:
    post:
      tags: [upload]
      summary: Confirm upload completion
      description: |
        Notify server that upload to B2 is complete. Triggers async processing job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [upload_id]
              properties:
                upload_id:
                  type: string
                  format: uuid
                etags:
                  type: array
                  items:
                    type: string
                  description: ETags from multipart upload (optional, for verification)
      responses:
        '202':
          description: Upload confirmed, processing queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  photo_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    example: "processing"
                  estimated_completion:
                    type: integer
                    example: 30
                    description: Estimated seconds until thumbnails ready
        '404':
          description: upload_id not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
...